move tests to their proper position
