+++ treehasher/defaulthasher/hasher_test.go
@@ -1,0 +1,107 @@
+package defaulthasher
+
+import (
+	"crypto"
+	_ "crypto/sha256"
+	"reflect"
+	"testing"
+)
+
+func Test_namespacedTreeHasher_HashLeaf(t *testing.T) {
+	zeroNID := []byte{0}
+	oneNID := []byte{1}
+	longNID := []byte("namespace")
+
+	defaultRawData := []byte("a blockchain is a chain of blocks")
+
+	// Note: ensure we only hash in the raw data without the namespace prefixes
+	emptyHash := sum(crypto.SHA256, []byte{LeafPrefix}, []byte{})
+	defaultHash := sum(crypto.SHA256, []byte{LeafPrefix}, defaultRawData)
+
+	oneNIDLeaf := append(oneNID, defaultRawData...)
+	longNIDLeaf := append(longNID, defaultRawData...)
+
+	tests := []struct {
+		name  string
+		nsLen int
+		leaf  []byte
+		want  []byte
+	}{
+		{"1 byte namespaced empty leaf", 1, zeroNID, append(append(zeroNID, zeroNID...), emptyHash...)},
+		{"1 byte namespaced empty leaf", 1, oneNID, append(append(oneNID, oneNID...), emptyHash...)},
+		{"1 byte namespaced leaf with data", 1, oneNIDLeaf, append(append(oneNID, oneNID...), defaultHash...)},
+		{"namespaced empty leaf", 9, longNIDLeaf, append(append(longNID, longNID...), defaultHash...)},
+		{"namespaced leaf with data", 9, longNID, append(append(longNID, longNID...), emptyHash...)},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			n := New(tt.nsLen, crypto.SHA256)
+			if got := n.HashLeaf(tt.leaf); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("HashLeaf() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_namespacedTreeHasher_HashNode(t *testing.T) {
+	sum(crypto.SHA256, []byte{NodePrefix}, []byte{0, 0, 0, 0}, []byte{1, 1, 1, 1})
+	type children struct {
+		l []byte
+		r []byte
+	}
+
+	tests := []struct {
+		name     string
+		nidLen   int
+		children children
+		want     []byte
+	}{
+		{"leftmin<rightmin && leftmax<rightmax", 2,
+			children{[]byte{0, 0, 0, 0}, []byte{1, 1, 1, 1}},
+			append(
+				[]byte{0, 0, 1, 1},
+				sum(crypto.SHA256, []byte{NodePrefix}, []byte{0, 0, 0, 0}, []byte{1, 1, 1, 1})...,
+			),
+		},
+		{"leftmin==rightmin && leftmax<rightmax", 2,
+			children{[]byte{0, 0, 0, 0}, []byte{0, 0, 1, 1}},
+			append(
+				[]byte{0, 0, 1, 1},
+				sum(crypto.SHA256, []byte{NodePrefix}, []byte{0, 0, 0, 0}, []byte{0, 0, 1, 1})...,
+			),
+		},
+		{"leftmin==rightmin && leftmax>rightmax", 2,
+			children{[]byte{0, 0, 1, 1}, []byte{0, 0, 0, 1}},
+			append(
+				[]byte{0, 0, 1, 1},
+				sum(crypto.SHA256, []byte{NodePrefix}, []byte{0, 0, 1, 1}, []byte{0, 0, 0, 1})...,
+			),
+		},
+		// XXX: can this happen in practice? or is this an invalid state?
+		{"leftmin>rightmin && leftmax<rightmax", 2,
+			children{[]byte{1, 1, 0, 0}, []byte{0, 0, 0, 1}},
+			append(
+				[]byte{0, 0, 0, 1},
+				sum(crypto.SHA256, []byte{NodePrefix}, []byte{1, 1, 0, 0}, []byte{0, 0, 0, 1})...,
+			),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			n := New(tt.nidLen, crypto.SHA256)
+			if got := n.HashNode(tt.children.l, tt.children.r); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("HashNode() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func sum(hash crypto.Hash, data ...[]byte) []byte {
+	h := hash.New()
+	for _, d := range data {
+		//nolint:errcheck
+		h.Write(d)
+	}
+
+	return h.Sum(nil)
+}
